refactor: 응원톡 내부 컴포넌트 구조 변경
