+++ apps/spectator/app/game/[id]/_components/CheerTalk/_components/Banner/Banner.css.ts
