+++ apps/spectator/app/game/[id]/_components/CheerTalk/_components/Banner/index.tsx
