+++ apps/spectator/app/game/[id]/_components/CheerTalk/_components/EntryButton/EntryButton.css.ts
