+++ apps/spectator/app/game/[id]/_components/CheerTalk/_components/Item/index.tsx
