+++ apps/spectator/app/game/[id]/_components/CheerTalk/_components/List/List.css.ts
